backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
argon2-cffi==21.3.0
python-dotenv==1.0.0
httpx==0.25.2
SQLAlchemy==2.0.23
slowapi==0.1.9
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0